+++ src/Pixel.hpp
@@ -5,25 +5,25 @@ class Pixel {
     public:
         Pixel();
 
-        Pixel(int r, int g, int b);
+        Pixel(double r, double g, double b);
 
         Pixel(const Pixel& other);
 
         Pixel& operator=(const Pixel& other);
 
-        void set_r(int r);
-        void set_g(int g);
-        void set_b(int b);
-        int get_r() const;
-        int get_g() const;
-        int get_b() const;
+        void set_r(double r);
+        void set_g(double g);
+        void set_b(double b);
+        double get_r() const;
+        double get_g() const;
+        double get_b() const;
 
 
 
     private:
-        int red;
-        int green;
-        int blue;
+        double red;
+        double green;
+        double blue;
 };
 
 #endif
